WIP some clean up
